Más cambios

+++ app/Http/Controllers/UserController.php
@@ -3,19 +3,33 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserFormRequest;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
+use Illuminate\Pagination\LengthAwarePaginator;
 use App\User;
+use App\Role;
 
 class UserController extends Controller
 {
+
+    private function verify(){
+        if(!Auth::user()){
+            return redirect('/');
+        }
+    }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $users = User::all();
+        if($request){
+            $query = trim($request->get('search'));
+            $users = User::where('name','LIKE','%'.$query.'%')->orderBy('id','asc')->paginate();
+            return view('usuarios.index',['users'=>$users,'search'=>$query]);
+        }
+        $users = User::paginate(15);
         return view('usuarios.index',['users'=>$users]);
     }
 
@@ -26,7 +40,8 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view("usuarios.create");
+        $this->verify();
+        return view("usuarios.create",["roles"=>Role::all()]);
     }
 
     /**
@@ -37,13 +52,15 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->verify();
         $usuario = new User();
 
         $usuario->name = request('name');
         $usuario->email = request('email');
-        $usuario->password = request('password');
+        $usuario->password = bcrypt(request('password'));
 
         $usuario->save();
+        $usuario->asignarRol(Role::findOrFail(request('rol')));
 
         return redirect('/usuarios');
     }
@@ -56,6 +73,7 @@ class UserController extends Controller
      */
     public function show($id)
     {
+        $this->verify();
         return view('usuarios.show',['user'=>User::findOrFail($id)]);
     }
 
@@ -67,7 +85,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        return view('usuarios.edit',['user'=>User::findOrFail($id)]);
+        $this->verify();
+        $user = User::findOrFail($id);
+        $role = new Role();
+        if(isset($user->tieneRole()[0])){
+            $role = $user->tieneRole()[0];
+        }
+        return view('usuarios.edit', ['user'=>$user,'role'=>$role, 'roles'=>Role::all()]);
     }
 
     /**
@@ -79,12 +103,20 @@ class UserController extends Controller
      */
     public function update(UserFormRequest $request, $id)
     {
+        $this->verify();
         $usuario = User::findOrFail($id);
 
         $usuario->name = $request->get('name');
         $usuario->email = $request->get('email');
 
         $usuario->update();
+        $role = $usuario->roles;
+        if(count($role)>0){
+            $role_id = $role[0]->id;
+            User::find($id)->roles()->updateExistingPivot($role_id, ['role_id'=>$request->get('rol') ]);
+        } else {
+            $usuario->asignarRol(Role::findOrFail(request('rol')));
+        }
 
         return redirect('/usuarios');
     }
@@ -97,6 +129,11 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->verify();
+        if(Auth::user()->id == $id){
+            return redirect('/usuarios');
+        }
+        $this->verify();
         $usuario = User::findOrFail($id);
         $usuario->delete();
